requirements.txt
requests
confluent-kafka
pycryptodome

SQLAlchemy